New fields for CarDealership - preferred cars params. All fixtures were moved to one file, all imports were remade.
