+++ car/tests/test_engine_http_methods.py
@@ -7,7 +7,7 @@ from rest_framework.test import APIClient
 
 from car.models import Engine
 from car.serializers.engine_serializers import EngineSerializer
-from car.tests.car_app_fixtures import engine
+from tests.project_fixtures import engine
 
 c = APIClient()
