+++ dealership/filters.py
@@ -2,8 +2,6 @@
 from django_countries.fields import CountryField
 from django_filters import rest_framework as filters
 
-# from django_countries import countries
-from car.models import Car
 from dealership.models import CarDealership
 
 
@@ -13,10 +11,7 @@ class CarDealershipFilter(filters.FilterSet):
     location = CountryField()
     balance = filters.RangeFilter()
     is_active = filters.BooleanFilter()
-    preferred_cars_list = filters.ChoiceFilter(
-        choices=[(car.id, car) for car in Car.objects.all()]
-    )
 
     class Meta:
         model = CarDealership
-        fields = ["location", "balance", "preferred_cars_list", "is_active"]
+        fields = ["location", "balance", "is_active"]
