+++ dealership/models.py
@@ -2,6 +2,8 @@
 from django.db import models
 from django_countries.fields import CountryField
 
+from car.models import CarTypes, TankTypes
+
 
 class CarDealership(models.Model):
     """
@@ -15,7 +17,28 @@ class CarDealership(models.Model):
         max_digits=10, decimal_places=2, verbose_name="Car dealership's balance"
     )
     preferred_cars_list = models.ManyToManyField(
-        "car.Car", verbose_name="List of preferred cars"
+        "car.Car", verbose_name="List of preferred cars", null=True, blank=True
+    )
+    preferred_car_release_year_from = models.IntegerField(
+        null=True, blank=True, verbose_name="Release year from"
+    )
+    preferred_car_release_year_to = models.IntegerField(
+        null=True, blank=True, verbose_name="Release year to"
+    )
+
+    preferred_car_type = models.CharField(
+        null=True,
+        blank=True,
+        choices=[(car_type.name, car_type.value) for car_type in CarTypes],
+        verbose_name="Car type",
+        max_length=20,
+    )
+    preferred_fuel_type = models.CharField(
+        null=True,
+        blank=True,
+        choices=[(fuel_type.name, fuel_type.value) for fuel_type in TankTypes],
+        verbose_name="Fuel type",
+        max_length=20,
     )
     is_active = models.BooleanField(verbose_name="Is active")
